work around s3 issues when building using our own packages
s3 does not handle special characters in objects URLs gracefully.
Furthermore, there seems to be a long-standing bug that causes
urls containing a '+' (%2b) to conflict with urls containing a ' '
(%20), resulting in the uploaded files only being found if uploaded
with a whitespace-literal instead of the plus-sign.
The workaround here is removing any '+' when uploading or fetching
from s3 via the gardenlinux-cache.